added cloud9 IDE to development tool
My company has a great number of Magento clients, so many that it'll be inefficient to use version control and copy the development environment onto local machine. And so we came up with a solution to use cloud9 IDE to edit code on the server directly from the cloud. It's so much nicer than using the PhpStorm - Filezilla duo.